specs: supervisor api
Take a quick pass at the supervisor api. This is useful because there
are many teams integrating into this right now.

op-geth has already integrated, the op-reth integration is ongoing as
well as the rbuilder integration should be happening soon.

This enables us to point to spec as the source of truth in how the API
is meant to work.